requirements.txt
boto3
aws_requests_auth
awscli
requests_aws4auth
